+++ lib/rucaptcha/view_helpers.rb
@@ -2,7 +2,11 @@ module RuCaptcha
   module ViewHelpers
     def rucaptcha_input_tag(opts = {})
       opts[:name] = '_rucaptcha'
-      opts[:type] = 'email'
+      opts[:type] = 'text'
+      opts[:autocorrect] = 'off'
+      opts[:autocapitalize] = 'off'
+      opts[:pattern] = '[0-9a-z]*'
+      opts[:maxlength] = RuCaptcha.config.len
       opts[:autocomplete] = 'off'
       tag(:input, opts)
     end